Change default number type size to long long

+++ thinglang/runtime/types/core/NumberType.h
@@ -25,7 +25,7 @@ class NumberInstance : public BaseThingInstance {
 
     
     /** Members **/
-    int val{};
+    long long val{};
     
     
 	~NumberInstance() override = default;
